Move conversion to to_dict function
